refactor: remove abstract varying socket class to simpluify the code
